requirements.txt
flask
torch
torchvision
Pillow
Flask
Flask-WTF
Flask-Login
Flask-SQLAlchemy
email-validator
python-dotenv
